Created a machine module
Created the machine module and moved everything StateMachine related to
the machine module.

Note: the imports ``from states.models import StateMachine`` will keep
working
